requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Multimodal AI and Computer Vision
openai>=1.10.0
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
moviepy>=1.0.3

# Audio processing
librosa>=0.10.0
speech-recognition>=3.10.0

# Web scraping and data processing
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0

# Optional: Hugging Face models
sentence-transformers>=2.2.0
diffusers>=0.24.0

# Google Gemini AI
google-generativeai>=0.3.0

# Anthropic (optional)
anthropic>=0.8.0 